Calculate the resilience values for Ising landscapes

The resilience is calculated based on the shape of the potential landscape and the prior knowledge about the qualitatively different parts of the system. Two resilience indicators are calculated separately, and their difference is used to represent a general resilience of the system in favor of the first phase. Within each phase, the potential difference between the local maximum and the local minimum (if multiple minimums exist, use the one that is further from the other phase; and the local maximum should always be on the side to the other phase) is used to represent the resilience of this phase.

Usage

calculate_resilience(l, ...)

# S3 method for `2d_Isingland`
calculate_resilience(l, split_value = 0.5 * l$Nvar, ...)

# S3 method for `2d_Isingland_matrix`
calculate_resilience(l, split_value = 0.5 * l$Nvar, ...)

Arguments

l

An Isingland object constructed with make_2d_Isingland() or make_2d_Isingland_matrix().

...

Not in use.

split_value

An integer to specify the number of active nodes used to split two resilience ranges. Default is half of the number of nodes.

Value

calculate_resilience.2d_Isingland()

Returns a calculate_resilience.2d_Isingland project, which contains the following elements:

dist

The distribution tibble which is the same as in the input l.

effective_minindex1,effective_maxindex1,effective_minindex2,effective_maxindex2

The (row)indices in dist that were used as the positions of the local minimums and maximums in two parts.

resilience1,resilience2,resilience_diff

The resilience measures for the first (left) part, the second part (right), and their difference.

calculate_resilience.2d_Isingland_matrix()

Returns a resilience_2d_Isingland_matrix object, which is a tibble containing columns of the varying parameters and a column resilience of the calculate_resilience.2d_Isingland objects for each landscape.

When print()ed, a verbal description of the resilience metrics is shown. Use the summary() method for a tidy version of the outputs.